@@ -21,14 +21,14 @@ import (
 	"errors"
 	"fmt"
 	"math"
-	"math/big"
-	"strings"
 	"sync"
+	"sync/atomic"
 	"time"
 
-	"github.com/gogo/protobuf/proto"
+	"google.golang.org/protobuf/proto"
 
 	pb "github.com/apache/pulsar-client-go/pulsar/internal/pulsar_proto"
+	"github.com/bits-and-blooms/bitset"
 )
 
 type messageID struct {
@@ -36,35 +36,34 @@ type messageID struct {
 	entryID      int64
 	batchIdx     int32
 	partitionIdx int32
+	batchSize    int32
 }
 
-var latestMessageID = messageID{
+var latestMessageID = &messageID{
 	ledgerID:     math.MaxInt64,
 	entryID:      math.MaxInt64,
 	batchIdx:     -1,
 	partitionIdx: -1,
+	batchSize:    0,
 }
 
-var earliestMessageID = messageID{
+var earliestMessageID = &messageID{
 	ledgerID:     -1,
 	entryID:      -1,
 	batchIdx:     -1,
 	partitionIdx: -1,
+	batchSize:    0,
 }
 
 type trackingMessageID struct {
-	messageID
+	*messageID
 
 	tracker      *ackTracker
 	consumer     acker
 	receivedTime time.Time
 }
 
-func (id trackingMessageID) Undefined() bool {
-	return id == trackingMessageID{}
-}
-
-func (id trackingMessageID) Ack() error {
+func (id *trackingMessageID) Ack() error {
 	if id.consumer == nil {
 		return errors.New("consumer is nil in trackingMessageID")
 	}
@@ -75,32 +74,62 @@ func (id trackingMessageID) Ack() error {
 	return nil
 }
 
-func (id trackingMessageID) Nack() {
+func (id *trackingMessageID) AckWithResponse() error {
+	if id.consumer == nil {
+		return errors.New("consumer is nil in trackingMessageID")
+	}
+	if id.ack() {
+		return id.consumer.AckIDWithResponse(id)
+	}
+
+	return nil
+}
+
+func (id *trackingMessageID) Nack() {
 	if id.consumer == nil {
 		return
 	}
 	id.consumer.NackID(id)
 }
 
-func (id trackingMessageID) NackByMsg(msg Message) {
+func (id *trackingMessageID) NackByMsg(msg Message) {
 	if id.consumer == nil {
 		return
 	}
 	id.consumer.NackMsg(msg)
 }
 
-func (id trackingMessageID) ack() bool {
+func (id *trackingMessageID) ack() bool {
 	if id.tracker != nil && id.batchIdx > -1 {
 		return id.tracker.ack(int(id.batchIdx))
 	}
 	return true
 }
 
-func (id messageID) isEntryIDValid() bool {
+func (id *trackingMessageID) ackCumulative() bool {
+	if id.tracker != nil && id.batchIdx > -1 {
+		return id.tracker.ackCumulative(int(id.batchIdx))
+	}
+	return true
+}
+
+func (id *trackingMessageID) prev() *trackingMessageID {
+	return &trackingMessageID{
+		messageID: &messageID{
+			ledgerID:     id.ledgerID,
+			entryID:      id.entryID - 1,
+			partitionIdx: id.partitionIdx,
+		},
+		tracker:  id.tracker,
+		consumer: id.consumer,
+	}
+}
+
+func (id *messageID) isEntryIDValid() bool {
 	return id.entryID >= 0
 }
 
-func (id messageID) greater(other messageID) bool {
+func (id *messageID) greater(other *messageID) bool {
 	if id.ledgerID != other.ledgerID {
 		return id.ledgerID > other.ledgerID
 	}
@@ -112,44 +141,49 @@ func (id messageID) greater(other messageID) bool {
 	return id.batchIdx > other.batchIdx
 }
 
-func (id messageID) equal(other messageID) bool {
+func (id *messageID) equal(other *messageID) bool {
 	return id.ledgerID == other.ledgerID &&
 		id.entryID == other.entryID &&
 		id.batchIdx == other.batchIdx
 }
 
-func (id messageID) greaterEqual(other messageID) bool {
+func (id *messageID) greaterEqual(other *messageID) bool {
 	return id.equal(other) || id.greater(other)
 }
 
-func (id messageID) Serialize() []byte {
+func (id *messageID) Serialize() []byte {
 	msgID := &pb.MessageIdData{
 		LedgerId:   proto.Uint64(uint64(id.ledgerID)),
 		EntryId:    proto.Uint64(uint64(id.entryID)),
 		BatchIndex: proto.Int32(id.batchIdx),
 		Partition:  proto.Int32(id.partitionIdx),
+		BatchSize:  proto.Int32(id.batchSize),
 	}
 	data, _ := proto.Marshal(msgID)
 	return data
 }
 
-func (id messageID) LedgerID() int64 {
+func (id *messageID) LedgerID() int64 {
 	return id.ledgerID
 }
 
-func (id messageID) EntryID() int64 {
+func (id *messageID) EntryID() int64 {
 	return id.entryID
 }
 
-func (id messageID) BatchIdx() int32 {
+func (id *messageID) BatchIdx() int32 {
 	return id.batchIdx
 }
 
-func (id messageID) PartitionIdx() int32 {
+func (id *messageID) PartitionIdx() int32 {
 	return id.partitionIdx
 }
 
-func (id messageID) String() string {
+func (id *messageID) BatchSize() int32 {
+	return id.batchSize
+}
+
+func (id *messageID) String() string {
 	return fmt.Sprintf("%d:%d:%d", id.ledgerID, id.entryID, id.partitionIdx)
 }
 
@@ -164,43 +198,66 @@ func deserializeMessageID(data []byte) (MessageID, error) {
 		int64(msgID.GetEntryId()),
 		msgID.GetBatchIndex(),
 		msgID.GetPartition(),
+		msgID.GetBatchSize(),
 	)
 	return id, nil
 }
 
-func newMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32) MessageID {
-	return messageID{
+func newMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32, batchSize int32) MessageID {
+	return &messageID{
 		ledgerID:     ledgerID,
 		entryID:      entryID,
 		batchIdx:     batchIdx,
 		partitionIdx: partitionIdx,
+		batchSize:    batchSize,
 	}
 }
 
-func newTrackingMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32,
-	tracker *ackTracker) trackingMessageID {
-	return trackingMessageID{
-		messageID: messageID{
+func fromMessageID(msgID MessageID) *messageID {
+	return &messageID{
+		ledgerID:     msgID.LedgerID(),
+		entryID:      msgID.EntryID(),
+		batchIdx:     msgID.BatchIdx(),
+		partitionIdx: msgID.PartitionIdx(),
+		batchSize:    msgID.BatchSize(),
+	}
+}
+
+func newTrackingMessageID(ledgerID int64, entryID int64, batchIdx int32, partitionIdx int32, batchSize int32,
+	tracker *ackTracker) *trackingMessageID {
+	return &trackingMessageID{
+		messageID: &messageID{
 			ledgerID:     ledgerID,
 			entryID:      entryID,
 			batchIdx:     batchIdx,
 			partitionIdx: partitionIdx,
+			batchSize:    batchSize,
 		},
 		tracker:      tracker,
 		receivedTime: time.Now(),
 	}
 }
 
-func toTrackingMessageID(msgID MessageID) (trackingMessageID, bool) {
-	if mid, ok := msgID.(messageID); ok {
-		return trackingMessageID{
-			messageID:    mid,
-			receivedTime: time.Now(),
-		}, true
-	} else if mid, ok := msgID.(trackingMessageID); ok {
-		return mid, true
-	} else {
-		return trackingMessageID{}, false
+// checkMessageIDType checks if the MessageID is user-defined
+func checkMessageIDType(msgID MessageID) (valid bool) {
+	switch msgID.(type) {
+	case *trackingMessageID:
+		return true
+	case *chunkMessageID:
+		return true
+	case *messageID:
+		return true
+	default:
+		return false
+	}
+}
+
+func toTrackingMessageID(msgID MessageID) (trackingMsgID *trackingMessageID) {
+	if mid, ok := msgID.(*trackingMessageID); ok {
+		return mid
+	}
+	return &trackingMessageID{
+		messageID: fromMessageID(msgID),
 	}
 }
 
@@ -243,6 +300,8 @@ type message struct {
 	replicatedFrom      string
 	redeliveryCount     uint32
 	schema              Schema
+	schemaVersion       []byte
+	schemaInfoCache     *schemaInfoCache
 	encryptionContext   *EncryptionContext
 	index               *uint64
 	brokerPublishTime   *time.Time
@@ -293,7 +352,18 @@ func (msg *message) GetReplicatedFrom() string {
 }
 
 func (msg *message) GetSchemaValue(v interface{}) error {
+	if msg.schemaVersion != nil {
+		schema, err := msg.schemaInfoCache.Get(msg.schemaVersion)
+		if err != nil {
+			return err
+		}
+		return schema.Decode(msg.payLoad, v)
+	}
 	return msg.schema.Decode(msg.payLoad, v)
+}
+
+func (msg *message) SchemaVersion() []byte {
+	return msg.schemaVersion
 }
 
 func (msg *message) ProducerName() string {
@@ -312,14 +382,14 @@ func (msg *message) BrokerPublishTime() *time.Time {
 	return msg.brokerPublishTime
 }
 
-func newAckTracker(size int) *ackTracker {
-	var batchIDs *big.Int
-	if size <= 64 {
-		shift := uint32(64 - size)
-		setBits := ^uint64(0) >> shift
-		batchIDs = new(big.Int).SetUint64(setBits)
-	} else {
-		batchIDs, _ = new(big.Int).SetString(strings.Repeat("1", size), 2)
+func (msg *message) size() int {
+	return len(msg.payLoad)
+}
+
+func newAckTracker(size uint) *ackTracker {
+	batchIDs := bitset.New(size)
+	for i := uint(0); i < size; i++ {
+		batchIDs.Set(i)
 	}
 	return &ackTracker{
 		size:     size,
@@ -329,8 +399,9 @@ func newAckTracker(size int) *ackTracker {
 
 type ackTracker struct {
 	sync.Mutex
-	size     int
-	batchIDs *big.Int
+	size           uint
+	batchIDs       *bitset.BitSet
+	prevBatchAcked uint32
 }
 
 func (t *ackTracker) ack(batchID int) bool {
@@ -339,12 +410,121 @@ func (t *ackTracker) ack(batchID int) bool {
 	}
 	t.Lock()
 	defer t.Unlock()
-	t.batchIDs = t.batchIDs.SetBit(t.batchIDs, batchID, 0)
-	return len(t.batchIDs.Bits()) == 0
+	t.batchIDs.Clear(uint(batchID))
+	return t.batchIDs.None()
+}
+
+func (t *ackTracker) ackCumulative(batchID int) bool {
+	if batchID < 0 {
+		return true
+	}
+	t.Lock()
+	defer t.Unlock()
+	for i := 0; i <= batchID; i++ {
+		t.batchIDs.Clear(uint(i))
+	}
+	return t.batchIDs.None()
+}
+
+func (t *ackTracker) hasPrevBatchAcked() bool {
+	return atomic.LoadUint32(&t.prevBatchAcked) == 1
+}
+
+func (t *ackTracker) setPrevBatchAcked() {
+	atomic.StoreUint32(&t.prevBatchAcked, 1)
 }
 
 func (t *ackTracker) completed() bool {
 	t.Lock()
 	defer t.Unlock()
-	return len(t.batchIDs.Bits()) == 0
+	return t.batchIDs.None()
+}
+
+func (t *ackTracker) toAckSet() []int64 {
+	t.Lock()
+	defer t.Unlock()
+	if t.batchIDs.None() {
+		return nil
+	}
+	bytes := t.batchIDs.Bytes()
+	ackSet := make([]int64, len(bytes))
+	for i := 0; i < len(bytes); i++ {
+		ackSet[i] = int64(bytes[i])
+	}
+	return ackSet
+}
+
+type chunkMessageID struct {
+	*messageID
+
+	firstChunkID *messageID
+	receivedTime time.Time
+
+	consumer acker
+}
+
+func newChunkMessageID(firstChunkID *messageID, lastChunkID *messageID) *chunkMessageID {
+	return &chunkMessageID{
+		messageID:    lastChunkID,
+		firstChunkID: firstChunkID,
+		receivedTime: time.Now(),
+	}
+}
+
+func (id chunkMessageID) String() string {
+	return fmt.Sprintf("%s;%s", id.firstChunkID.String(), id.messageID.String())
+}
+
+func (id chunkMessageID) Serialize() []byte {
+	msgID := &pb.MessageIdData{
+		LedgerId:   proto.Uint64(uint64(id.ledgerID)),
+		EntryId:    proto.Uint64(uint64(id.entryID)),
+		BatchIndex: proto.Int32(id.batchIdx),
+		Partition:  proto.Int32(id.partitionIdx),
+		FirstChunkMessageId: &pb.MessageIdData{
+			LedgerId:   proto.Uint64(uint64(id.firstChunkID.ledgerID)),
+			EntryId:    proto.Uint64(uint64(id.firstChunkID.entryID)),
+			BatchIndex: proto.Int32(id.firstChunkID.batchIdx),
+			Partition:  proto.Int32(id.firstChunkID.partitionIdx),
+		},
+	}
+	data, _ := proto.Marshal(msgID)
+	return data
+}
+
+type topicMessageID struct {
+	track *trackingMessageID
+	topic string
+}
+
+func (t *topicMessageID) Serialize() []byte {
+	return t.track.Serialize()
+}
+
+func (t *topicMessageID) LedgerID() int64 {
+	return t.track.LedgerID()
+}
+
+func (t *topicMessageID) EntryID() int64 {
+	return t.track.EntryID()
+}
+
+func (t *topicMessageID) BatchIdx() int32 {
+	return t.track.BatchIdx()
+}
+
+func (t *topicMessageID) PartitionIdx() int32 {
+	return t.track.PartitionIdx()
+}
+
+func (t *topicMessageID) BatchSize() int32 {
+	return t.track.BatchSize()
+}
+
+func (t *topicMessageID) String() string {
+	return t.track.String()
+}
+
+func (t *topicMessageID) Topic() string {
+	return t.topic
 }
