@@ -84,7 +84,7 @@ func (h *keyBasedBatches) Val(key string) *batchContainer {
 // NewKeyBasedBatchBuilder init batch builder and return BatchBuilder
 // pointer. Build a new key based batch message container.
 func NewKeyBasedBatchBuilder(
-	maxMessages uint, maxBatchSize uint, producerName string, producerID uint64,
+	maxMessages uint, maxBatchSize uint, maxMessageSize uint32, producerName string, producerID uint64,
 	compressionType pb.CompressionType, level compression.Level,
 	bufferPool BuffersPool, logger log.Logger, encryptor crypto.Encryptor,
 ) (BatchBuilder, error) {
@@ -92,7 +92,7 @@ func NewKeyBasedBatchBuilder(
 	bb := &keyBasedBatchContainer{
 		batches: newKeyBasedBatches(),
 		batchContainer: newBatchContainer(
-			maxMessages, maxBatchSize, producerName, producerID,
+			maxMessages, maxBatchSize, maxMessageSize, producerName, producerID,
 			compressionType, level, bufferPool, logger, encryptor,
 		),
 		compressionType: compressionType,
@@ -131,6 +131,10 @@ func (bc *keyBasedBatchContainer) Add(
 	metadata *pb.SingleMessageMetadata, sequenceIDGenerator *uint64,
 	payload []byte,
 	callback interface{}, replicateTo []string, deliverAt time.Time,
+	schemaVersion []byte, multiSchemaEnabled bool,
+	useTxn bool,
+	mostSigBits uint64,
+	leastSigBits uint64,
 ) bool {
 	if replicateTo != nil && bc.numMessages != 0 {
 		// If the current batch is not empty and we're trying to set the replication clusters,
@@ -150,7 +154,7 @@ func (bc *keyBasedBatchContainer) Add(
 	if batchPart == nil {
 		// create batchContainer for new key
 		t := newBatchContainer(
-			bc.maxMessages, bc.maxBatchSize, bc.producerName, bc.producerID,
+			bc.maxMessages, bc.maxBatchSize, bc.maxMessageSize, bc.producerName, bc.producerID,
 			bc.compressionType, bc.level, bc.buffersPool, bc.log, bc.encryptor,
 		)
 		batchPart = &t
@@ -158,10 +162,14 @@ func (bc *keyBasedBatchContainer) Add(
 	}
 
 	// add message to batch container
-	batchPart.Add(
+	add := batchPart.Add(
 		metadata, sequenceIDGenerator, payload, callback, replicateTo,
 		deliverAt,
+		schemaVersion, multiSchemaEnabled, useTxn, mostSigBits, leastSigBits,
 	)
+	if !add {
+		return false
+	}
 	addSingleMessageToBatch(bc.buffer, metadata, payload)
 
 	bc.numMessages++
@@ -183,51 +191,35 @@ func (bc *keyBasedBatchContainer) reset() {
 	bc.batches.containers = map[string]*batchContainer{}
 }
 
-// Flush all the messages buffered in multiple batches and wait until all
+// FlushBatches all the messages buffered in multiple batches and wait until all
 // messages have been successfully persisted.
-func (bc *keyBasedBatchContainer) FlushBatches() (
-	batchesData []Buffer, sequenceIDs []uint64, callbacks [][]interface{}, errors []error,
-) {
+func (bc *keyBasedBatchContainer) FlushBatches() []*FlushBatch {
 	if bc.numMessages == 0 {
 		// No-Op for empty batch
-		return nil, nil, nil, nil
+		return nil
 	}
 
 	bc.log.Debug("keyBasedBatchContainer flush: messages: ", bc.numMessages)
-	var batchesLen = len(bc.batches.containers)
-	var idx = 0
-	sortedKeys := make([]string, 0, batchesLen)
-
-	batchesData = make([]Buffer, batchesLen)
-	sequenceIDs = make([]uint64, batchesLen)
-	callbacks = make([][]interface{}, batchesLen)
-	errors = make([]error, batchesLen)
-
 	bc.batches.l.RLock()
 	defer bc.batches.l.RUnlock()
-	for k := range bc.batches.containers {
-		sortedKeys = append(sortedKeys, k)
-	}
-	sort.Strings(sortedKeys)
-	for _, k := range sortedKeys {
-		container := bc.batches.containers[k]
-		b, s, c, err := container.Flush()
-		if b != nil {
-			batchesData[idx] = b
-			sequenceIDs[idx] = s
-			callbacks[idx] = c
-			errors[idx] = err
+	flushBatches := make([]*FlushBatch, 0, len(bc.batches.containers))
+	for _, container := range bc.batches.containers {
+		batch := container.Flush()
+		if batch == nil {
+			continue
 		}
-		idx++
+		flushBatches = append(flushBatches, batch)
 	}
 
 	bc.reset()
-	return batchesData, sequenceIDs, callbacks, errors
+	// Sort items by sequenceID
+	sort.Slice(flushBatches, func(i, j int) bool {
+		return flushBatches[i].SequenceID < flushBatches[j].SequenceID
+	})
+	return flushBatches
 }
 
-func (bc *keyBasedBatchContainer) Flush() (
-	batchData Buffer, sequenceID uint64, callbacks []interface{}, err error,
-) {
+func (bc *keyBasedBatchContainer) Flush() *FlushBatch {
 	panic("multi batches container not support Flush(), please use FlushBatches() instead")
 }
 
