@@ -129,7 +129,9 @@ func (tv *TableViewImpl) partitionUpdateCheck() error {
 				if err != nil {
 					tv.logger.Errorf("read next message failed for %s: %w", partition, err)
 				}
-				tv.handleMessage(msg)
+				if msg != nil {
+					tv.handleMessage(msg)
+				}
 			}
 			ctx, cancelFunc := context.WithCancel(context.Background())
 			tv.cancelRaders[partition] = cancelReader{
@@ -245,13 +247,18 @@ func (tv *TableViewImpl) handleMessage(msg Message) {
 	tv.dataMu.Lock()
 	defer tv.dataMu.Unlock()
 
-	payload := reflect.Indirect(reflect.New(tv.options.SchemaValueType)).Interface()
-	if err := msg.GetSchemaValue(&payload); err != nil {
-		tv.logger.Errorf("msg.GetSchemaValue() failed with %w; msg is %v", msg, err)
+	payload := reflect.New(tv.options.SchemaValueType)
+	if len(msg.Payload()) == 0 {
+		delete(tv.data, msg.Key())
+	} else {
+		if err := msg.GetSchemaValue(payload.Interface()); err != nil {
+			tv.logger.Errorf("msg.GetSchemaValue() failed with %v; msg is %v", err, msg)
+		}
+		tv.data[msg.Key()] = reflect.Indirect(payload).Interface()
 	}
-	tv.data[msg.Key()] = payload
+
 	for _, listener := range tv.listeners {
-		if err := listener(msg.Key(), payload); err != nil {
+		if err := listener(msg.Key(), reflect.Indirect(payload).Interface()); err != nil {
 			tv.logger.Errorf("table view listener failed for %v: %w", msg, err)
 		}
 	}
@@ -263,9 +270,12 @@ func (tv *TableViewImpl) watchReaderForNewMessages(ctx context.Context, reader R
 		if err != nil {
 			tv.logger.Errorf("read next message failed for %s: %w", reader.Topic(), err)
 		}
-		if errors.Is(err, context.Canceled) {
+		var e *Error
+		if (errors.As(err, &e) && e.Result() == ConsumerClosed) || errors.Is(err, context.Canceled) {
 			return
 		}
-		tv.handleMessage(msg)
+		if msg != nil {
+			tv.handleMessage(msg)
+		}
 	}
 }
